feat: add savings goal milestone notification
